Revert "Fix : user_id 타입 수정 Post 속성 수정"
This reverts commit b3a04cc35e5bfd9cc2c1e46aefa627ec4d3b05cf.

+++ TigerCave/src/main/java/com/sparta/tigercave/repository/UsersRepository.java
@@ -3,8 +3,9 @@ package com.sparta.tigercave.repository;
 import com.sparta.tigercave.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<User, Long> {
+public interface UsersRepository extends JpaRepository<User, BigInteger> {
     Optional<User> findByUsername(String username);
 }
